// src/app/models/pelicula.ts
export interface Pelicula {
    id: string;
    title: string;
    
    original_title: string;
    original_title_romanised: string;
    image: string;
    movie_banner:string;
    release_date: Int16Array;
    running_time: Int16Array;
    rt_score:Int16Array;
    people: string;
    species: string;
    locations: string;
    vehicles: string;
    url:string

    director: string;
    year: string;
    description: string;
    
}
        